Fixed SOCKET_read bug to work with single quotes
Earlier, the function would crashed while trying to load web pages with
a lot of content. This is because often, pages with a lot of content may
have single quotes somewhere in the page. Basically, not only do
double-quotes have to be escaped by double double-quotes in order to be
parsed as valid Scilab strings, even single-quotes have to be escaped by
double single-quotes.